Adds sspd arrest records in csv format
This is a csv that can be loaded into excel or a database for further analysis.  The arrestee_race column data has been cleaned up to standardize on current formats.  For example, 'Blk', 'Black', and 'BLACK' are now all 'BLACK/AFRICAN AMERICAN'.  The data was extracted from a FOIL request that resulted in scanned PDFs, which in turn were run through OCR to convert the data to fields.